services/idv/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.10.0
onnxruntime>=1.18.0
insightface>=0.7.3
pyzbar>=0.1.9
